Frontend: Alternate window resize system that eliminates reconnects.
Instead of sizing the RDP control to the size of the client area, size it to the size of the screen and tell the backend to resize its browser window to the visible width and height. This works because the browser window lives at screen offset (0,0) which is the top-left of the RDP control and therefore the top-left of the client area. Resizing is now much more smooth (still not 100% perfect due to the use of a Timer to limit resize events) and no longer prone to reconnection errors. In fact, there is now no need to ever disconnect and reconnect RDP at all.