- Implemented "Find in Page" feature in frontend and backend. Tested with Wikipedia. Worked straight away. - Added vbAccelerator's CommonDialog6 classes with source code and as a separate project/ActiveX DLL (with changed GUID) to prepare for file upload/file download common dialog prompts - Removed COMCTL32.OCX dependency since it's no longer needed due to the above - Added vbAccelerator's cFindReplace class into the CommonDialog6 project as well - Patched vbAccelerator's cFindReplace class to also provide a DialogClosed event
Mr. McMahon figured out all of the intricate details of making the FindText common dialog API work with VB6, which must have taken days of trial and error. Take a look at mFindReplaceHook.bas to see the lunacy required to coax VB6 into somehow doing the right thing. But the important thing is that it works.